refactor macro
