+++ bothan-api/server/src/config/manager/forex_info/sources.rs
@@ -4,6 +4,8 @@
 
 use serde::{Deserialize, Serialize};
 
+use crate::config::manager::band_serde::de_band_named;
+
 /// Configuration for the worker sources for forex asset info.
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct ForexSourceConfigs {
@@ -25,23 +27,6 @@ pub struct ForexSourceConfigs {
     /// will have an empty/default name. The custom deserializer `de_xenops` reconstructs the options
     #[serde(default, deserialize_with = "de_xenops")]
     pub band_xenops: Option<bothan_band::WorkerOpts>,
-}
-
-// Macro to generate deserialization functions for Band workers with preset names.
-// This macro defines a function that:
-// - Deserializes an Option<WorkerOpts>,
-// - If present, creates a new WorkerOpts with the given name and original URL/update_interval.
-macro_rules! de_band_named {
-    ($fn_name:ident, $name:expr) => {
-        fn $fn_name<'de, D>(d: D) -> Result<Option<bothan_band::WorkerOpts>, D::Error>
-        where
-            D: serde::Deserializer<'de>,
-        {
-            let v = Option::<bothan_band::WorkerOpts>::deserialize(d)?;
-            let v = v.map(|w| bothan_band::WorkerOpts::new($name, &w.url, Some(w.update_interval)));
-            Ok(v)
-        }
-    };
 }
 
 const BAND1_WORKER_NAME: &str = "band/owlet";
